requirements.txt
alembic==1.7.3
click==8.0.1
dataclasses==0.8
Flask==2.0.1
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.8.1
importlib-resources==5.2.2
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.5
MarkupSafe==2.0.1
migrate==0.3.8
pkg-resources==0.0.0
psycopg2-binary==2.8.6
SQLAlchemy==1.4.25
typing-extensions==3.10.0.2
Werkzeug==2.0.1
zipp==3.6.0
phonenumbers==8.12.33
SQLAlchemy-Utils==0.37.8
Flask-WTF==0.15.1
WTForms~=2.3.3
enum34~=1.1.6
idna~=2.6
requests~=2.18.4
cryptography~=2.1.4
pip~=21.2.4
ipaddress~=1.0.17
python-dotenv~=0.18.0
blinker~=1.4
python-dateutil~=2.8.2
configparser~=4.0.2
Django~=1.11.29
setuptools~=58.1.0
six~=1.16.0
Pillow~=5.1.0
flask-bootstrap==3.3.7.1
Flask-Email==1.4.4
Flask-Mail==0.9.1
Flask-Bcrypt==0.7.1
